Unify editors to use one Triggers editor for all count queries
