+++ src/datasource/components/QueryEditor/TriggersQueryEditor.tsx
@@ -1,14 +1,20 @@
 import _ from 'lodash';
-import React, { useEffect } from 'react';
+import React, { useEffect, FormEvent } from 'react';
 import { useAsyncFn } from 'react-use';
 
 import { SelectableValue } from '@grafana/data';
-import { InlineField, InlineSwitch, Select } from '@grafana/ui';
+import { InlineField, InlineSwitch, Input, Select } from '@grafana/ui';
 import { QueryEditorRow } from './QueryEditorRow';
 import { MetricPicker } from '../../../components';
 import { getVariableOptions } from './utils';
 import { ZabbixDatasource } from '../../datasource';
 import { ZabbixMetricsQuery } from '../../types';
+
+const countByOptions: Array<SelectableValue<string>> = [
+  { value: '', label: 'All triggers' },
+  { value: 'problems', label: 'Problems' },
+  { value: 'items', label: 'Items' },
+];
 
 const severityOptions: Array<SelectableValue<number>> = [
   { value: 0, label: 'Not classified' },
@@ -77,9 +83,55 @@ export const TriggersQueryEditor = ({ query, datasource, onChange }: Props) => {
     return options;
   }, [query.group.filter, query.host.filter]);
 
+  const loadProxyOptions = async () => {
+    const proxies = await datasource.zabbix.getProxies();
+    const options = proxies?.map((proxy) => ({
+      value: proxy.host,
+      label: proxy.host,
+    }));
+    options.unshift(...getVariableOptions());
+    return options;
+  };
+
+  const [{ loading: proxiesLoading, value: proxiesOptions }, fetchProxies] = useAsyncFn(async () => {
+    const options = await loadProxyOptions();
+    return options;
+  }, []);
+
+  const loadItemOptions = async (group: string, host: string, app: string, itemTag: string) => {
+    const groupFilter = datasource.replaceTemplateVars(group);
+    const hostFilter = datasource.replaceTemplateVars(host);
+    const appFilter = datasource.replaceTemplateVars(app);
+    const tagFilter = datasource.replaceTemplateVars(itemTag);
+    const options = {
+      itemtype: 'num',
+      showDisabledItems: query.options.showDisabledItems,
+    };
+    const items = await datasource.zabbix.getAllItems(groupFilter, hostFilter, appFilter, tagFilter, options);
+    let itemOptions: Array<SelectableValue<string>> = items?.map((item) => ({
+      value: item.name,
+      label: item.name,
+    }));
+    itemOptions = _.uniqBy(itemOptions, (o) => o.value);
+    itemOptions.unshift(...getVariableOptions());
+    return itemOptions;
+  };
+
+  const [{ loading: itemsLoading, value: itemOptions }, fetchItems] = useAsyncFn(async () => {
+    const options = await loadItemOptions(
+      query.group.filter,
+      query.host.filter,
+      query.application.filter,
+      query.itemTag.filter
+    );
+    return options;
+  }, [query.group.filter, query.host.filter, query.application.filter, query.itemTag.filter]);
+
   // Update suggestions on every metric change
   const groupFilter = datasource.replaceTemplateVars(query.group?.filter);
   const hostFilter = datasource.replaceTemplateVars(query.host?.filter);
+  const appFilter = datasource.replaceTemplateVars(query.application?.filter);
+  const tagFilter = datasource.replaceTemplateVars(query.itemTag?.filter);
 
   useEffect(() => {
     fetchGroups();
@@ -92,6 +144,23 @@ export const TriggersQueryEditor = ({ query, datasource, onChange }: Props) => {
   useEffect(() => {
     fetchApps();
   }, [groupFilter, hostFilter]);
+
+  useEffect(() => {
+    fetchProxies();
+  }, []);
+
+  useEffect(() => {
+    fetchItems();
+  }, [groupFilter, hostFilter, appFilter, tagFilter]);
+
+  const onTextFilterChange = (prop: string) => {
+    return (v: FormEvent<HTMLInputElement>) => {
+      const newValue = v?.currentTarget?.value;
+      if (newValue !== null) {
+        onChange({ ...query, [prop]: { filter: newValue } });
+      }
+    };
+  };
 
   const onFilterChange = (prop: string) => {
     return (value: string) => {
@@ -107,8 +176,25 @@ export const TriggersQueryEditor = ({ query, datasource, onChange }: Props) => {
     }
   };
 
+  const onCountByChange = (option: SelectableValue) => {
+    if (option.value !== null) {
+      onChange({ ...query, countTriggersBy: option.value! });
+    }
+  };
+
   return (
     <>
+      <QueryEditorRow>
+        <InlineField label="Count by" labelWidth={12}>
+          <Select
+            isSearchable={false}
+            width={24}
+            value={query.countTriggersBy}
+            options={countByOptions}
+            onChange={onCountByChange}
+          />
+        </InlineField>
+      </QueryEditorRow>
       <QueryEditorRow>
         <InlineField label="Group" labelWidth={12}>
           <MetricPicker
@@ -128,6 +214,17 @@ export const TriggersQueryEditor = ({ query, datasource, onChange }: Props) => {
             onChange={onFilterChange('host')}
           />
         </InlineField>
+        {query.countTriggersBy === 'problems' && (
+          <InlineField label="Proxy" labelWidth={12}>
+            <MetricPicker
+              width={24}
+              value={query.proxy?.filter}
+              options={proxiesOptions}
+              isLoading={proxiesLoading}
+              onChange={onFilterChange('proxy')}
+            />
+          </InlineField>
+        )}
       </QueryEditorRow>
       <QueryEditorRow>
         <InlineField label="Application" labelWidth={12}>
@@ -139,19 +236,52 @@ export const TriggersQueryEditor = ({ query, datasource, onChange }: Props) => {
             onChange={onFilterChange('application')}
           />
         </InlineField>
+        {query.countTriggersBy === 'problems' && (
+          <>
+            <InlineField label="Problem" labelWidth={12}>
+              <Input
+                width={24}
+                defaultValue={query.trigger?.filter}
+                placeholder="Problem name"
+                onBlur={onTextFilterChange('trigger')}
+              />
+            </InlineField>
+            <InlineField label="Tags" labelWidth={12}>
+              <Input
+                width={24}
+                defaultValue={query.tags?.filter}
+                placeholder="tag1:value1, tag2:value2"
+                onBlur={onTextFilterChange('tags')}
+              />
+            </InlineField>
+          </>
+        )}
+        {query.countTriggersBy === 'items' && (
+          <InlineField label="Item" labelWidth={12}>
+            <MetricPicker
+              width={24}
+              value={query.item.filter}
+              options={itemOptions}
+              isLoading={itemsLoading}
+              onChange={onFilterChange('item')}
+            />
+          </InlineField>
+        )}
+      </QueryEditorRow>
+      <QueryEditorRow>
         <InlineField label="Min severity" labelWidth={12}>
           <Select
             isSearchable={false}
             width={24}
-            value={query.triggers?.minSeverity}
+            value={query.options?.minSeverity}
             options={severityOptions}
             onChange={onMinSeverityChange}
           />
         </InlineField>
         <InlineField label="Count" labelWidth={12}>
           <InlineSwitch
-            value={query.triggers?.count}
-            onChange={() => onChange({ ...query, triggers: { ...query.triggers, count: !query.triggers?.count } })}
+            value={query.options?.count}
+            onChange={() => onChange({ ...query, options: { ...query.options, count: !query.options?.count } })}
           />
         </InlineField>
       </QueryEditorRow>
