+++ src/datasource/constants.ts
@@ -13,9 +13,7 @@ export const MODE_TEXT = '2';
 export const MODE_ITEMID = '3';
 export const MODE_TRIGGERS = '4';
 export const MODE_PROBLEMS = '5';
-export const MODE_TRIGGERS_ITEM = '6';
-export const MODE_TRIGGERS_PROBLEM = '7';
-export const MODE_MACROS = '8';
+export const MODE_MACROS = '6';
 
 // Triggers severity
 export const SEV_NOT_CLASSIFIED = 0;
